All Iulian's work as one commit
We don't want to merge this into master because we want to keep your
history. But GitHub makes it basically impossible to review an entire
branch of changes at once. So pushing this to a sidebranch as one commit
just makes it easy to leave review comments. You'll want to make changes
on your main branch, as usual.